--- Basecode.Data/Dtos/Address/AddressCreationDto.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Basecode.Data.Dtos.Address
-{
-    public class AddressCreationDto
-    {
-    }
-}
