--- Basecode.Data/Dtos/Address/AddressUpdationDto.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Basecode.Data.Dtos.Address
-{
-    public class AddressUpdationDto
-    {
-        public int Id { get; set; }
-        public string? Street { get; set; }
-        public string? City { get; set; }
-        public Type? Province { get; set; }
-        public Type? ZipCode { get; set; }
-    }
-}
