--- Basecode.Data/Dtos/CharacterReferences/CharacterReferencesCreationDto.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Basecode.Data.Dtos.CharacterReferences
-{
-    public class CharacterReferencesCreationDto
-    {
-        public int Id { get; set; }
-        public int ApplicantId { get; set; }
-        public string Name { get; set; } = string.Empty;
-        public string Relationship { get; set; } = string.Empty;
-        public string Email { get; set; } = string.Empty;
-        public string MobileNumber { get; set; } = string.Empty;
-    }
-}
