+++ Basecode.Data/Dtos/CharacterReferencesCreationDto.cs
@@ -4,11 +4,12 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Basecode.Data.Dtos.CharacterReferences
+namespace Basecode.Data.Dtos
 {
-    public class CharacterReferencesUpdationDto
+    public class CharacterReferencesCreationDto
     {
         public int Id { get; set; }
+        public int ApplicantId { get; set; }
         public string Name { get; set; } = string.Empty;
         public string Relationship { get; set; } = string.Empty;
         public string Email { get; set; } = string.Empty;
