+++ Basecode.Data/Repositories/ApplicantRepository.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Net;
 using System.Text;
 using System.Threading.Tasks;
 using Basecode.Data.Interfaces;
@@ -16,9 +17,25 @@ namespace Basecode.Data.Repositories
             _context = context;
         }
 
+        public void Add(Applicant applicant)
+        {
+            _context.Applicants.Add(applicant);
+            _context.SaveChanges();
+        }
+
         public Applicant GetById(int id)
         {
-            return _context.Applicant.Find(id);
+            return _context.Applicants.Find(id);
+        }
+
+        public Applicant GetByName(string name)
+        {
+            return _context.Applicants.FirstOrDefault(e => e.Name == name);
+        }
+
+        public IQueryable<Applicant> RetrieveAll()
+        {
+            return this.GetDbSet<Applicant>();
         }
     }
 }
