+++ Basecode.Services/Interfaces/IAddressService.cs
@@ -4,7 +4,6 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Basecode.Data.Dtos;
-using Basecode.Data.Dtos.Address;
 using Basecode.Data.Models;
 
 namespace Basecode.Services.Interfaces
@@ -15,7 +14,5 @@ namespace Basecode.Services.Interfaces
         Address GetByCity(string city);
         void Add(AddressCreationDto address);
         Address GetById(int id);
-        void Update(AddressUpdationDto address);
-        void Delete(int id);
     }
 }
