+++ Basecode.Services/Interfaces/ICharacterReferencesService.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using Basecode.Data.Dtos.CharacterReferences;
+using Basecode.Data.Dtos;
 using Basecode.Data.Models;
 
 namespace Basecode.Services.Interfaces
@@ -11,10 +11,8 @@ namespace Basecode.Services.Interfaces
     public interface ICharacterReferencesService
     {
         List<CharacterReferences> RetrieveAll();
-        CharacterReferences GetByCity(string city);
+        CharacterReferences GetByName(string name);
         void Add(CharacterReferencesCreationDto characterReferences);
         CharacterReferences GetById(int id);
-        void Update(CharacterReferencesUpdationDto characterReferences);
-        void Delete(int id);
     }
 }
